ml-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.1
prophet==1.1.6
requests==2.32.3
python-dotenv==1.0.1
